AHP and Tau fix
AHP can now only be found within 5 ms after peak time, this is to make sure that AHPs of different AHP are always the fast component. Often this fast component can only be found in the first 1-2 AP's in a train since then the slow AHP component becomes more dominant. The slow AHP component are also saved and plotted with the plotanalysis function